% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdjhdohis.R
\name{rd_update_his_main}
\alias{rd_update_his_main}
\title{Title:历史票据更新，}
\usage{
rd_update_his_main(input_excel, token = "F91CF3E3-8962-47F2-823F-C5CCAAFC66CA")
}
\arguments{
\item{jh_test_token，口令}{}
}
\value{
更新结果
}
\description{
Title:历史票据更新，
}
\examples{
rd_update_his_main(input_excel="需处理历史单据清单.xlsx",token='F91CF3E3-8962-47F2-823F-C5CCAAFC66CA')

}
